Moved some classes from modVM to fastVM
- IPrecompiledContract is now PrecompiledContract and is now in
  modPrecompiled. IContractFactory is removed. ExecutionContext and
  SideEffects have been moved from modVM into the fvm.
