+++ modMcf/src/org/aion/mcf/vm/types/KernelInterfaceForFastVM.java
@@ -1,4 +1,4 @@
-package org.aion.vm;
+package org.aion.mcf.vm.types;
 
 import java.math.BigInteger;
 import org.aion.base.db.IRepositoryCache;
@@ -6,7 +6,6 @@ import org.aion.base.util.ByteArrayWrapper;
 import org.aion.mcf.core.AccountState;
 import org.aion.mcf.db.IBlockStoreBase;
 import org.aion.mcf.valid.TxNrgRule;
-import org.aion.mcf.vm.types.DataWord;
 import org.aion.vm.api.interfaces.Address;
 import org.aion.vm.api.interfaces.KernelInterface;
