+++ modPrecompiled/src/module-info.java
@@ -3,7 +3,6 @@ module aion.precompiled {
     requires aion.mcf;
     requires aion.base;
     requires aion.crypto;
-    requires aion.vm;
     requires slf4j.api;
     requires jsr305;
     requires commons.collections4;
