+++ modPrecompiled/test/org/aion/precompiled/contracts/ATB/BridgeTestUtils.java
@@ -4,7 +4,7 @@ import org.aion.base.type.AionAddress;
 import org.aion.crypto.AddressSpecs;
 import org.aion.crypto.HashUtil;
 import org.aion.mcf.vm.types.DataWord;
-import org.aion.vm.ExecutionContext;
+import org.aion.fastvm.ExecutionContext;
 
 public class BridgeTestUtils {
     static ExecutionContext dummyContext() {
