+++ modPrecompiled/test/org/aion/precompiled/contracts/BenchmarkTest.java
@@ -8,8 +8,8 @@ import org.aion.base.type.AionAddress;
 import org.aion.mcf.config.CfgFork;
 import org.aion.mcf.vm.types.DataWord;
 import org.aion.precompiled.ContractFactory;
-import org.aion.vm.ExecutionContext;
-import org.aion.vm.IPrecompiledContract;
+import org.aion.fastvm.ExecutionContext;
+import org.aion.precompiled.type.PrecompiledContract;
 import org.aion.zero.impl.config.CfgAion;
 import org.apache.commons.lang3.RandomUtils;
 import org.junit.After;
@@ -97,7 +97,7 @@ public class BenchmarkTest {
                         blockNrgLimit,
                         blockDifficulty);
 
-        IPrecompiledContract ct;
+        PrecompiledContract ct;
         // warm up
         for (int i = 0; i < WARMUP; i++) {
             ct = cf.getPrecompiledContract(ctx, null);
@@ -134,7 +134,7 @@ public class BenchmarkTest {
         //                blockNrgLimit,
         //                blockDifficulty);
         //
-        //        IPrecompiledContract ct;
+        //        PrecompiledContract ct;
         //        // warm up
         //        for (int i = 0; i < WARMUP; i++) {
         //            ct = cf.getPrecompiledContract(ctx, null);
