+++ modPrecompiled/test/org/aion/precompiled/contracts/EDVerifyContractTest.java
@@ -20,8 +20,8 @@ import org.aion.mcf.config.CfgFork;
 import org.aion.mcf.vm.types.DataWord;
 import org.aion.precompiled.ContractFactory;
 import org.aion.precompiled.PrecompiledResultCode;
-import org.aion.vm.ExecutionContext;
-import org.aion.vm.IPrecompiledContract;
+import org.aion.fastvm.ExecutionContext;
+import org.aion.precompiled.type.PrecompiledContract;
 import org.aion.vm.api.interfaces.TransactionResult;
 import org.aion.zero.impl.config.CfgAion;
 import org.apache.commons.lang3.RandomUtils;
@@ -105,7 +105,7 @@ public class EDVerifyContractTest {
                         blockTimestamp,
                         blockNrgLimit,
                         blockDifficulty);
-        IPrecompiledContract contract = new ContractFactory().getPrecompiledContract(ctx, null);
+        PrecompiledContract contract = new ContractFactory().getPrecompiledContract(ctx, null);
 
         assertNotNull(contract);
         TransactionResult result = contract.execute(input, 21000L);
@@ -135,7 +135,7 @@ public class EDVerifyContractTest {
                         blockTimestamp,
                         blockNrgLimit,
                         blockDifficulty);
-        IPrecompiledContract contract = new ContractFactory().getPrecompiledContract(ctx, null);
+        PrecompiledContract contract = new ContractFactory().getPrecompiledContract(ctx, null);
 
         assertNotNull(contract);
         TransactionResult result = contract.execute(input, 21000L);
@@ -169,7 +169,7 @@ public class EDVerifyContractTest {
                         blockTimestamp,
                         blockNrgLimit,
                         blockDifficulty);
-        IPrecompiledContract contract = new ContractFactory().getPrecompiledContract(ctx, null);
+        PrecompiledContract contract = new ContractFactory().getPrecompiledContract(ctx, null);
 
         assertNotNull(contract);
         TransactionResult result = contract.execute(input, 21000L);
@@ -200,7 +200,7 @@ public class EDVerifyContractTest {
                         blockTimestamp,
                         blockNrgLimit,
                         blockDifficulty);
-        IPrecompiledContract contract = new ContractFactory().getPrecompiledContract(ctx, null);
+        PrecompiledContract contract = new ContractFactory().getPrecompiledContract(ctx, null);
 
         TransactionResult result = contract.execute(input, 2999L);
         assertThat(result.getResultCode().toInt()).isEqualTo(PrecompiledResultCode.OUT_OF_NRG.toInt());
@@ -229,7 +229,7 @@ public class EDVerifyContractTest {
                         blockTimestamp,
                         blockNrgLimit,
                         blockDifficulty);
-        IPrecompiledContract contract = new ContractFactory().getPrecompiledContract(ctx, null);
+        PrecompiledContract contract = new ContractFactory().getPrecompiledContract(ctx, null);
 
         TransactionResult result = contract.execute(input, 21000L);
