--- modVM/src/org/aion/vm/IContractFactory.java
@@ -1,12 +1,0 @@
-package org.aion.vm;
-
-import org.aion.vm.api.interfaces.TransactionContext;
-
-public interface IContractFactory {
-
-    // Using FastVM interface for now... until full solution in place.
-
-    IPrecompiledContract getPrecompiledContract(
-            TransactionContext context,
-            KernelInterfaceForFastVM track);
-}
